Refactor Animations into separate module, take care of some warnings

+++ ProjectRedemptionEngine/src/modules/rendering/model/rendermaterial.cpp
@@ -106,7 +106,7 @@ namespace PRE
 				_impl.pShaderProgram->SetMat4(
 					"PRE_BONETRANSFORMS",
 					&boneTransforms[0],
-					boneTransforms.size()
+					(unsigned int)boneTransforms.size()
 				);
 			}
 		}
